后台管理模块开发，我的聊天室模块开发，登录权限控制

+++ src/components/UserPerson.vue
@@ -15,6 +15,9 @@
             <span
               style="font-size:24px;color:#fff;margin-left: 15px;"
             >{{userData && userData.nickname ||'未知用户，请登录'}}</span>
+            <span
+              style="font-size:12px;color:#999;margin: 10px 15px;"
+            >权限：{{permissionMap[userData && userData.permission ||'0']}}</span>
             <div>
               <el-badge :value="0" class="item">
                 <el-button size="small">私信</el-button>
@@ -25,6 +28,21 @@
               <el-badge :value="messageList &&messageList.length || 0" class="item">
                 <el-button size="small">通知</el-button>
               </el-badge>
+              <el-button
+                style="margin-left:15px;"
+                size="small"
+                v-if="userInfo && userInfo.permission =='0'"
+                @click="verifyUser"
+                type="danger"
+              >申请认证</el-button>
+              <router-link to="/management">
+                <el-button
+                  style="margin-left:15px;"
+                  size="small"
+                  v-if="userInfo && userInfo.permission =='1'"
+                  type="primary"
+                >后台管理 >></el-button>
+              </router-link>
             </div>
           </div>
         </div>
@@ -77,9 +95,12 @@
                 <el-menu-item index="3-2">收藏视频</el-menu-item>
               </router-link>
             </el-submenu>
-            <el-submenu v-if="userInfo.permission =='1' || userInfo.permission =='2'" index="4">
+            <el-submenu
+              v-if="userInfo && userInfo.permission =='1' || userInfo && userInfo.permission =='2'"
+              index="4"
+            >
               <template slot="title">
-                <i class="el-icon-user"></i>
+                <i class="el-icon-s-flag"></i>
                 <span>我的活动&比赛</span>
               </template>
               <router-link to="/person/add_activity">
@@ -89,15 +110,27 @@
                 <el-menu-item index="4-2">我的活动&比赛</el-menu-item>
               </router-link>
             </el-submenu>
-            <el-submenu index="5">
+            <el-submenu v-if="userInfo && userInfo.permission =='２'" index="5">
+              <template slot="title">
+                <i class="el-icon-s-comment"></i>
+                <span>我的聊天室</span>
+              </template>
+              <router-link to="/person/add_room">
+                <el-menu-item index="5-1">新建聊天室</el-menu-item>
+              </router-link>
+              <router-link to="/person/my_room">
+                <el-menu-item index="5-2">我的聊天室</el-menu-item>
+              </router-link>
+            </el-submenu>
+            <el-submenu index="6">
               <template slot="title">
                 <i class="el-icon-user"></i>
                 <span>我的账户</span>
               </template>
               <router-link to="/person/update_user_info">
-                <el-menu-item index="5-1">修改个人信息</el-menu-item>
+                <el-menu-item index="6-1">修改个人信息</el-menu-item>
               </router-link>
-              <el-menu-item @click="logout" index="5-2">退出登录</el-menu-item>
+              <el-menu-item @click="logout" index="6-2">退出登录</el-menu-item>
             </el-submenu>
           </el-menu>
         </el-col>
@@ -106,17 +139,63 @@
         </el-col>
       </div>
     </el-row>
+    <el-dialog title="申请认证" :loading="dialoading" :visible.sync="dialogVisible" width="60%">
+      <div style="padding:20px">
+        <el-form
+          :rules="updateUserRules"
+          ref="updateUserForm"
+          :label-position="'top'"
+          :model="updateUserForm"
+          label-width="80px"
+        >
+          <el-form-item label="手机号:" prop="phone">
+            <el-input placeholder="请输入11位手机号" maxlength="11" v-model="updateUserForm.phone"></el-input>
+          </el-form-item>
+          <el-form-item label="个人介绍:" prop="introduce">
+            <el-input placeholder="请输入个人介绍" v-model="updateUserForm.introduce"></el-input>
+          </el-form-item>
+          <el-form-item style="width:100%;">
+            <div style="margin-top:80px">
+              <el-button
+                type="danger"
+                icon="el-icon-s-promotion"
+                @click="submit('updateUserForm')"
+              >提交</el-button>
+            </div>
+          </el-form-item>
+        </el-form>
+      </div>
+    </el-dialog>
   </div>
 </template>
 
 <script>
 import { mapState, mapMutations, mapActions } from "vuex";
+import { UpdateUserInfo } from "../api/user_api";
+import { permissionMap } from "../utils/tools";
 export default {
   name: "UserPerson",
   data() {
     return {
+      dialoading: false,
+      dialogVisible: false,
+      updateUserForm: {
+        phone: "",
+        introduce: ""
+      },
+      updateUserContent: "",
+      updateUserRules: {
+        phone: [
+          { required: true, message: "请输入手机号", trigger: "blur" },
+          { min: 11, max: 11, message: "请输入正确的手机号", trigger: "blur" }
+        ],
+        introduce: [
+          { required: true, message: "请输入个人介绍", trigger: "blur" }
+        ]
+      },
       collectcounts: 0,
-      userData: {}
+      userData: {},
+      permissionMap
     };
   },
   created() {},
@@ -134,17 +213,71 @@ export default {
       "changeUserInfo"
     ]),
     ...mapActions([
+      "GetUserInfoById",
       "GetAllDanceSortList",
       "GetAllArticleList",
       "GetAllVideoList",
       "GetAllRotationImgList"
     ]),
+    //更新房间
+    submit(formname) {
+      this.$refs[formname].validate(valid => {
+        if (valid) {
+          if (this.updateUserForm.imgurl == "") {
+            this.$message.error("请上传房间海报");
+            return;
+          }
+          this.$confirm("确定提交吗？")
+            .then(_ => {
+              this.updateUserForm = {
+                ...this.updateUserForm,
+                permission: "6",
+                _id: this.userid
+              };
+              
+              UpdateUserInfo(this.updateUserForm)
+                .then(res => {
+                  
+                  this.updateUserForm = {
+                    phone: "",
+                    introduce: ""
+                  };
+                  this.fileList = [];
+                  this.content = "";
+                  this.$message.success("申请认证成功");
+                  this.dialogVisible = false;
+                  this.GetUserInfoById({ _id: this.userid });
+                })
+                .catch(err => {
+                  
+                  this.$message.error("出现错误，请稍候再试");
+                  this.dialogVisible = false;
+                });
+            })
+            .catch(_ => {
+              return;
+            });
+        } else {
+          return false;
+        }
+      });
+    },
+    verifyUser() {
+      this.dialoading = true;
+      
+      this.dialogVisible = true;
+      this.dialoading = false;
+    },
     logout() {
-      localStorage.setItem("accessToken", "");
-      this.changeIsLogin(false);
-      this.changeUserInfo({});
-      this.changeUserId("");
-      this.$router.push("/");
+      this.$confirm("确定退出登录吗？")
+        .then(_ => {
+          localStorage.setItem("accessToken", "");
+          this.changeIsLogin(false);
+          this.changeUserInfo({});
+          this.changeUserId("");
+          this.$router.push("/");
+        })
+        .catch(_ => {});
     }
   },
   computed: {
@@ -166,7 +299,15 @@ export default {
   watch: {
     userInfo: {
       handler(newval, old) {
-        console.log(newval);
+        
+        if (newval && newval.permission === "4") {
+          this.$message.error("您已被封禁，恢复帐号请联系网站管理员");
+          localStorage.setItem("accessToken", "");
+          this.changeIsLogin(false);
+          this.changeUserInfo({});
+          this.changeUserId("");
+          this.$router.push("/");
+        }
         this.userData = newval;
       },
       deep: true
